requirements.txt
asgiref==3.8.1
Django==5.1.1
django-modeltranslation==0.19.8
pillow==10.4.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
sqlparse==0.5.1
tzdata==2024.1
